% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config_csv_files.R
\name{create_config_csv_files}
\alias{create_config_csv_files}
\title{Create Configuration CSV Files for Reading and Writing Data}
\usage{
create_config_csv_files(
  user_config_path = "user_config.yml",
  output_dir_path = "config_files"
)
}
\arguments{
\item{user_config_path}{A character string specifying the path to the user-specified
\code{config.yml} file. Defaults to "user_config.yml".}

\item{output_dir_path}{A character string specifying the directory path where the configuration
CSV files will be written. Defaults to "config_files".}
}
\value{
NULL. The function writes two CSV files to the specified directory: \code{config_read.csv}
for reading configurations and \code{config_write.csv} for writing configurations.
}
\description{
This function initializes project configuration by creating two CSV files:
one for reading data and another for writing data.
It uses either a user-specified \code{config.yml} file or the package's default \code{config.yml} file
to determine the structure of these configuration files.
}
\examples{
\dontrun{
# Using the default config.yml
create_config_csv_files()
}

\dontrun{
# Specifying a user-defined config.yml
create_config_csv_files(user_config_path = "path/to/user_config.yml")
}

}
