% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_dot_arguments.R
\name{overwrite_dot_arguments}
\alias{overwrite_dot_arguments}
\title{Overwrite arguments with dot arguments}
\usage{
overwrite_dot_arguments(function_args, ...)
}
\arguments{
\item{function_args}{A named list of function arguments.}

\item{...}{Additional arguments passed with the dot (...). These arguments will
overwrite the corresponding ones in \code{function_args} if they have the same names.}
}
\value{
A list containing the original \code{function_args} with any overwritten values
from the dot arguments.
}
\description{
This function takes a list of function arguments and additional arguments passed
with the dot (...). It overwrites the function arguments with the values from the
dot arguments if they have the same names.
}
\examples{
# Example usage
original_args <- list(a = 1, b = 2)
result <- overwrite_dot_arguments(original_args, b = 3, c = 4)
result

}
