% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvprojector_write_package_config_yml.R
\name{vvprojector_write_package_config_yml}
\alias{vvprojector_write_package_config_yml}
\title{Write Package Config.yml to User-Defined Location}
\usage{
vvprojector_write_package_config_yml(output_path = getwd())
}
\arguments{
\item{output_path}{A character string specifying the path where the \code{config.yml} file should be written. Defaults to the current working directory.}
}
\value{
NULL. The function writes the \code{config.yml} file to the specified location.
}
\description{
This function retrieves the package's default \code{config.yml} file and writes it to a specified location.
If no location is specified, it defaults to the current working directory.
}
\examples{
# Write to the current working directory
\dontrun{
vvprojector_write_package_config_yml()
}

# Write to a specific location
\dontrun{
vvprojector_write_package_config_yml(output_path = "path/to/config/directory")
}

}
